Change to GET request
It appears USPS wrapped the POST request parameters that worked in the prior version in a USER_HISTORY_LIST: object; this version makes things work again. This version may be less capable than the prior one for handling multiple tracking numbers, but that is not the primary intended use case.